refactor: auth 모듈 패키지 구조 개선 (FeignClient 도입 대응)
