+++ auth/src/main/java/com/mossy/boundedContext/app/RefreshTokenUseCase.java
@@ -3,7 +3,7 @@ package com.mossy.boundedContext.app;
 import com.mossy.exception.DomainException;
 import com.mossy.exception.ErrorCode;
 import com.mossy.boundedContext.global.jwt.JwtProperties;
-import com.mossy.boundedContext.out.RefreshTokenRepository;
+import com.mossy.boundedContext.out.repository.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
